% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breach_events.R
\docType{data}
\name{breach_events}
\alias{breach_events}
\title{Simulation of breach event counts from Fig.5 of "COVID-19 in low-tolerance border quarantine systems: Impact of the Delta variant of SARS-CoV-2" using Poisson distribution}
\format{
\subsection{A data frame with 84 rows and 4 columns}{

\describe{
\item{day}{Days infectious in community}
\item{panel}{
\enumerate{
\item{A: Ancestral strain without vaccination}
\item{B: Ancestral strain with vaccination}
\item{C: Delta strain without vaccination}
\item{D: Delta strain with vaccination (increased transmissibility and decreased vaccine efficacy)}
}
}
\item{traveler_count_pois}{Traveler breach counts, generated by rpois() function}
\item{worker_count_pois}{Worker breach counts, generated by rpois() function}
}
}
}
\source{
\url{https://https://www.science.org/doi/10.1126/sciadv.abm3624#F5}
}
\usage{
breach_events
}
\description{
Simulation of breach event counts from Fig.5 of "COVID-19 in low-tolerance border quarantine systems: Impact of the Delta variant of SARS-CoV-2" using Poisson distribution
}
\keyword{datasets}
