% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_density_plots.R
\name{draw_density_plots}
\alias{draw_density_plots}
\title{Draw 4 density plots of a numeric column from a data frame by 4 different scenarios}
\usage{
draw_density_plots(col)
}
\arguments{
\item{col}{either "travelers" or "workers"; will draw density plots for either travelers breach or workers breach}
}
\value{
a ggplot2 object with geom density
}
\description{
Draw 4 density plots of a numeric column from a data frame by 4 different scenarios
}
\examples{
draw_density_plots("travelers")

}
