% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_histogram.R
\name{draw_histograms}
\alias{draw_histograms}
\title{Draw 4 histogram of a numeric column from a data frame by 4 different scenarios}
\usage{
draw_histograms(col)
}
\arguments{
\item{col}{either "travelers" or "workers"; will draw histograms for either travelers breach or workers breach}
}
\value{
a ggplot2 object with geom histogram
}
\description{
Draw 4 histogram of a numeric column from a data frame by 4 different scenarios
}
\examples{
draw_histograms("travelers")

}
