% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny_app.R
\name{run_shiny_app}
\alias{run_shiny_app}
\title{Launch ashiny app that allows users to interact with histograms, density plots, and data tables}
\usage{
run_shiny_app()
}
\description{
Launch ashiny app that allows users to interact with histograms, density plots, and data tables
}
